requirements.txt
colorzero==2.0
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
kiwisolver==1.4.5
lgpio==0.0.0.2
matplotlib==3.8.0
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
picamera==1.13
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
RPi.GPIO==0.7.1
six==1.16.0
tflite-runtime==2.14.0
